auto-fill-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-cloud-storage==2.10.0
langchain-google-genai==0.0.5
langchain-core==0.1.10
pypdf==3.17.4
python-docx==1.1.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
pydantic==2.5.0
requests==2.31.0 
werkzeug
